refactor: move IndyProofUtils logic to IndyProofFormatService. Make name and version also optional
