+++ packages/core/src/modules/proofs/formats/indy/index.ts
@@ -2,4 +2,3 @@ export * from './errors'
 export * from './models'
 export * from './IndyProofFormat'
 export * from './IndyProofFormatsServiceOptions'
-export * from './IndyProofUtils'
